backend/requirements/base.txt
Django==3.1.7
mysqlclient==2.0.3
pytest-django==4.1.0
pytest-cov==2.11.1
pytest-xdist==2.2.1